% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidaPlots.R
\name{BetweenViewBiplot}
\alias{BetweenViewBiplot}
\title{Biplots for Discriminant Scores or Canonical Correlation Variates between pairs of views}
\usage{
BetweenViewBiplot(
  object,
  Y,
  Xtest = NULL,
  color.palette = NULL,
  keep.loadings = NULL,
  plotIt = TRUE
)
}
\arguments{
\item{object}{the output from SIDA, SIDANet, and SELPCCA methods}

\item{Y}{a vector of class membership for grouping canonical correlatoin variates and discriminant scores.}

\item{color.palette}{character vector of length K (number of classes), specifying the colors to use for the classes, respectively.
Defaults to shades of blue and orange (color.BlueOrange). Other option includes red and green combinations (color.GreenRed)}

\item{keep.loadings}{numeric vector of length D (number of views), specifying how many variables
to represent on loadings plot for each view. This is useful
in situations where the number of variables selected is large, and could clutter the plot. If this number is more
than the variables selected, it will be set to the maximum number of variables selected for each view.
Default is plotting all selected variables.}

\item{A}{list of D entries containing test data. If not null, scores for biplots will be constructed for testing data.}
}
\value{
\item{NULL}{}
}
\description{
Biplots  to visualize discriminant scores/ canonical variates between pairs of views. It shows how
selected variables from the first and second discriminant (for SIDA and SIDANet)
or canonical correlation (for SELPCCA) vectors in a view is related to selected variables in another view.
Variables farther from the origin and close to first or second axis
have higher impact on first or second discriminant/canonical vectors, respectively.
Variables farther from the origin and between both first and second axes have similar higher contributions to the
first and second discriminant/canonical correlation vectors. In both situations, for SIDA and SIDANet, this suggests that
these variables contribute more to the separation of classes and association of views. For SELPCCA, this suggests that
these variables contribute more to the association between the two views. This plot can
only be generated for classification and association problems with 3 or more classes (SIDA and SIDANet),
or for CCA problems with two or more canonical correlation vectors requested (i.e. ncancorr > 1 for SELPCCA).
This plot shows the scores and loadings from pairs of views together. The scores are the
sum of scores for each view. Solid and dashed lines represent vectors for Views 1 and 2,
respectively.
}
\details{
The function will return loading plots, one for each view.
}
\examples{
data("sidanetData")
Xdata <- sidanetData[[1]]
Y <- sidanetData[[2]] #class membership already coded as 1,2,...
Xtestdata <- sidanetData[[3]]
Ytest <- sidanetData[[4]] #class membership already coded as 1,2,...
#edge information
myedges=sidanetData[[5]]
myedgeweight=sidanetData[[6]]
##---- call cross validation
mycvsidanet=cvSIDANet(Xdata,Y,myedges,myedgeweight,withCov=FALSE,plotIt=FALSE,Xtestdata=Xtestdata,
                     Ytest=Ytest)
BetweenViewBiplot(mycvsidanet, Y,keep.loadings=c(3,3) )
}
\references{
Sandra E. Safo, Eun Jeong Min, and Lillian Haine (2023), Sparse Linear Discriminant
Analysis for Multi-view Structured Data, Biometrics.
Sandra E. Safo, Jeongyoun Ahn, Yongho Jeon, and Sungkyu Jung (2018),
 Sparse Generalized Eigenvalue Problem with Application to Canonical
 Correlation Analysis for Integrative Analysis of Methylation and Gene
 Expression Data. Biometrics
}
\seealso{
\code{\link{cvSIDA}} \code{\link{DiscriminantPlots}}
\code{\link{CorrelationPlots}}
}
