% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVIDData}
\alias{COVIDData}
\title{Multiomics data pertaining to COVID-19}
\format{
A list with 3 elements:
\describe{
  \item{\code{COVIDData$Proteomics}}{Proteomics data. A data frame of size \eqn{120 \times 264}.
  Rows are samples and columns are variables.}
  \item{\code{COVIDData$RNAseq}}{RNASeq data. A data frame of size \eqn{120 \times 5800}.
  Rows are samples and columns are variables.}
  \item{\code{COVIDData$Clinical}}{Clinical and demographic data. A data frame of size \eqn{120 \times 18}.
  Rows are samples and columns are variables.}
}
}
\usage{
COVIDData
}
\description{
RNA Sequencing (RNASeq) and Proteomics data pertaining to COVID-19. Clinical data
are also available. Please refer to Overmyer et.al (2021) for a description of the data and Lipman et.al (2022) for how data
were pre-processed.
}
\references{
Multi-omic analysis reveals enriched pathways
associated with COVID-19 and COVID-19 severity. PLOS ONE, 17(4)
Overmyer, K.A., Shishkova, E., Miller, I.J., Balnis, J., Bernstein, M.N., Peters-Clarke, T.M., Meyer, J.G., Quan,
Q., Muehlbauer, L.K., Trujillo, E.A., et al.: Large-scale multi-omic analysis of covid-19 severity. Cell systems
12(1), 23–40 (2021)
}
\keyword{datasets}
