% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTables.R
\name{devianceTable}
\alias{devianceTable}
\title{Deviance Table for predicted, supervised data.}
\usage{
devianceTable(fit, Xtestdata1, Xtestdata2, Ytest)
}
\arguments{
\item{fit}{the output from the filter_supervised() function}

\item{Xtestdata1}{a matrix or data.frame with the first view's test X data}

\item{Xtestdata2}{a matrix or data.frame with the second view's test X data}

\item{Ytest}{a vector of test response values}
}
\value{
A data.frame
}
\description{
Produced a deviance table for supervised predictions.
}
