% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filter.unsupervised}
\alias{filter.unsupervised}
\title{Unsupervised Filtering}
\usage{
filter.unsupervised(
  X,
  method = "variance",
  pct.keep = 10,
  center = FALSE,
  scale = FALSE,
  standardize = FALSE,
  log2TransForm = FALSE,
  Xtest = NULL
)
}
\arguments{
\item{X}{A list containing all data sources. Each row must represent a
subject and each column represents a feature.}

\item{method}{Options are "variance" which will keep the \code{pct.keep} percent of
features with the highest variance, and "IQR", which will keep the features with the
median amount of variance (+/- pct.keep/2). Default is "variance".}

\item{pct.keep}{Percent of variables to keep in each view of data. Default is 10\%.}

\item{center}{Boolean on whether or not to center the features after filtering.}

\item{scale}{Boolean on whether or not to scale the features after filtering.}

\item{standardize}{Boolean on whether or not to center and scale the features to have mean 0 and variance 1 after filtering.}

\item{log2TransForm}{Boolean on whether or not to log2 transform the features prior to filtering. Will return an error if TRUE but data
have negative values.}

\item{Xtest}{Optional list containing test data. If included, filtering will be performed
only on the training data, X, but Xtest will be subsetted to the same group of features.}
}
\value{
A list containing the following
  \item{X}{List of the filtered X data}
  \item{Xtest}{List of the subsetted Xtest data}
  \item{method}{Method used for filtering}
  \item{var.mat}{Dataset containing the calculated mean and variances for each feature.}
}
\description{
Performs univariate unsupervised filtering on multi-source
 data. A separate model will be fit for each feature within each view of data
 and all features with p-values less than the specified threshold will be retained.
}
\examples{
##---- read in data
data(sidaData)

Xdata=sidaData[[1]]

data.red=filter.unsupervised(Xdata,  method="variance", pct.keep=10,
                    center=FALSE, scale=FALSE, standardize=FALSE,  log2TransForm=FALSE,
                     Xtest=NULL)
}
