% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{filter_supervised}
\alias{filter_supervised}
\title{Supervised Filtering}
\usage{
filter_supervised(
  X,
  Y,
  method = "linear",
  padjust = FALSE,
  adjmethod = "BH",
  thresh = 0.05,
  center = FALSE,
  scale = FALSE,
  standardize = FALSE,
  log2TransForm = FALSE,
  Xtest = NULL
)
}
\arguments{
\item{X}{A list containing all data sources. Each row must represent a
subject and each column represents a feature.}

\item{Y}{An outcome vector of length equal to the number of rows in each view of X.}

\item{method}{Options are "linear" for linear regression, "logistic" for logistic
regression, "t.test" for a 2-sample unpaired T-test, or "kw" for a Kruskal-Wallis
test. Default is "linear".}

\item{padjust}{Boolean on whether or not to adjust pvalue for multiple testing.
Default is "F".}

\item{adjmethod}{Options are "holm", "hochberg",  "hommel", "bonferroni", "BH"
"BY","fdr","none". Default is "BH" if padjust is True.}

\item{thresh}{P-value threshold to determine which features to keep after filtering.
Default will keep all features with a p-value < 0.05.}

\item{center}{Boolean on whether or not to center the features prior to filtering.}

\item{scale}{Boolean on whether or not to scale the features to have variance 1 prior to filtering.}

\item{standardize}{Boolean on whether or not to center and scale the features to have mean 0 and variance 1 prior to filtering.}

\item{log2TransForm}{Boolean on whether or not to log2 transform the features prior to filtering. Will return an error if TRUE but data
have negative values.}

\item{Xtest}{Optional list containing test data. If included, filtering will be performed
only on the training data, X, but Xtest will be subsetted to the same group of features.}
}
\value{
A list containing the following (and others):
  \item{X}{List of the filtered X data}
  \item{Y}{Vector of the outcome}
  \item{Xtest}{List of the subsetted Xtest data}
  \item{method}{Method used for filtering}
  \item{pval.mat}{Dataset containing the calculated p-values for each feature, coefficients, and whether significant.}
}
\description{
Performs univariate supervised filtering on multi-source
 data. A separate model will be fit for each feature within each view of data
 and all features with p-values less than the specified threshold will be retained.
}
\examples{
##---- read in data
data(sidaData)

Xdata=sidaData[[1]]
Y=sidaData[[2]]

data.red=filter_supervised(Xdata, Y,  method="t.test", padjust=FALSE,adjmethod=NULL, thresh=0.05,
center=FALSE, scale=FALSE, standardize=FALSE, log2TransForm=FALSE, Xtest=NULL)

##-----Plot Result via UMAP
umapPlot(data.red)
}
