% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidaPlots.R
\name{Loadings}
\alias{Loadings}
\title{Loadings}
\usage{
Loadings(fit)
}
\arguments{
\item{fit}{the output from cvSIDA, cvSIDANet, or cvselpcca methods}
}
\value{
A data.frame with three columns. The first two columns are the non-zero loadings
The last column is the View from which the loadings were obtained.
}
\description{
function to get loadings for variables selected
by SIDA, SIDANet, and SELPCCA methods.
}
