% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidaPlots.R
\name{LoadingsPlots}
\alias{LoadingsPlots}
\title{Loadings Plots}
\usage{
LoadingsPlots(object, color.line = "darkgray", keep.loadings = NULL)
}
\arguments{
\item{object}{the output from SIDA, SIDANet, and SELPCCA methods}

\item{color.line}{color to use for plotting direction vectors. Default is "darkgray".}

\item{keep.loadings}{numeric, specifying how many variables to represent on loadings plot. This is useful
in situations where the number of variables selected is large, and could clutter the plot. If this number is more
than the variables selected, it will be set to the maximum number of variables selected for each view.
Default is plotting all selected variables.}
}
\value{
\item{NULL}{}
}
\description{
Plots discriminant and canonical vectors  to visualize how
selected variables contribute to the first and second discriminant (for SIDA and SIDANet)
or canonical correlation (for SELPCCA) vectors.  Variables farther from the origin and close to first or second axis
have higher impact on first or second discriminant/canonical vectors, respectively.
Variables farther from the origin and between both first and second axes have similar higher contributions to the
first and second discriminant/canonical correlation vectors. In both situations, for SIDA and SIDANet, this suggests that
these variables contribute more to the separation of classes and association of views. For SELPCCA, this suggests that
these variables contribute more to the association between the two views. This plot can
only be generated for classification and association problems with 3 or more classes (SIDA and SIDANet),
or for CCA problems with two or more canonical correlation vectors requested (i.e. ncancorr > 1 for SELPCCA).
}
\details{
The function will return loading plots, one for each view.
}
\examples{
data("sidanetData")
Xdata <- sidanetData[[1]]
Y <- sidanetData[[2]] #class membership already coded as 1,2,...
Xtestdata <- sidanetData[[3]]
Ytest <- sidanetData[[4]] #class membership already coded as 1,2,...
#edge information
myedges=sidanetData[[5]]
myedgeweight=sidanetData[[6]]
##---- call cross validation
mycvsidanet=cvSIDANet(Xdata,Y,myedges,myedgeweight,withCov=FALSE,plotIt=FALSE,Xtestdata=Xtestdata,
                     Ytest=Ytest)
LoadingsPlots(mycvsidanet,keep.loadings=c(3,3))
}
\references{
Sandra E. Safo, Eun Jeong Min, and Lillian Haine (2023), Sparse Linear Discriminant
Analysis for Multi-view Structured Data, Biometrics.
Sandra E. Safo, Jeongyoun Ahn, Yongho Jeon, and Sungkyu Jung (2018),
 Sparse Generalized Eigenvalue Problem with Application to Canonical
 Correlation Analysis for Integrative Analysis of Methylation and Gene
 Expression Data. Biometrics
}
\seealso{
\code{\link{cvSIDA}} \code{\link{DiscriminantPlots}}
\code{\link{CorrelationPlots}}
}
