% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selp.R
\name{multiplescca}
\alias{multiplescca}
\title{Sparse canonical correlation vectors for fixed tuning parameters}
\usage{
multiplescca(
  Xdata1 = Xdata1,
  Xdata2 = Xdata2,
  ncancorr = ncancorr,
  Tau = Tau,
  CovStructure = "Iden",
  standardize = TRUE,
  maxiteration = 20,
  thresh = 1e-04
)
}
\arguments{
\item{Xdata1}{A matrix of size \eqn{n \times p} for first dataset.
Rows are samples and columns are variables.}

\item{Xdata2}{A matrix of size \eqn{n \times q} for second dataset.
Rows are samples and columns are variables.}

\item{ncancorr}{Number of canonical correlation vectors. Default is 1.}

\item{Tau}{A vector of matrix of fixed tuning parameters for each dataset.}

\item{CovStructure}{Covariance structure to use in estimating sparse
canonical correlation vectors. Either "Iden" or "Ridge". Iden assumes
the covariance matrix for each dataset is identity. Ridge uses the
sample covariance for each dataset. See reference article for more details.}

\item{standardize}{TRUE or FALSE. If TRUE, data will be normalized to have
mean zero and variance one for each variable. Default is TRUE.}

\item{maxiteration}{Maximum iteration for the algorithm if not converged.
Default is 20.}

\item{thresh}{Threshold for convergence. Default is 0.0001.}
}
\value{
The output is a list containing the following components.
\item{hatalpha}{Estimated sparse canonical correlation vectors for first dataset.}
\item{hatbeta}{Estimated sparse canonical correlation vectors for second dataset.}
\item{maxcorr}{Estimated canonical correlation coefficient.}
}
\description{
Function for estimating canonical correlation vectors for
a fixed tuning parameters for each dataset.
}
\details{
The function will return several R objects, which can be assigned
 to a variable. To see the results, use the “$" operator.
}
\examples{
\dontrun{
##---- read in data
data(selpData)

Xdata1=selpData[[1]]
Xdata2=selpData[[2]]

  ##---- estimate first canonical correlation vectors
ncancorr=1

#use cvtunerange for range of tuning parameters
mytunerange=cvtunerange(Xdata1=Xdata1,Xdata2=Xdata2,ncancorr=ncancorr,
                        CovStructure="Iden",standardize=TRUE)
print(mytunerange)

#Fix Tau for first and second datasets as 1.1 and 1.0 respectively
Tau=matrix(c(1,1.2,1),nrow=1)
mysparsevectors=multiplescca(Xdata1=Xdata1,Xdata2=Xdata2,ncancorr=ncancorr,
                             Tau=Tau, CovStructure="Iden",standardize=TRUE,
                             maxiteration=20, thresh=0.0001)

#example with two canonical correlation vectors
#use cvselpscca to obtain optimal tuning parameters
mycv=cvselpscca(Xdata1=Xdata1,Xdata2=Xdata2,ncancorr=ncancorr,
                CovStructure="Iden",isParallel=TRUE,ncores=NULL,nfolds=5,
                ngrid=10, standardize=TRUE,thresh=0.0001,maxiteration=20)

Tau=mycv$optTau
mysparsevectors=multiplescca(Xdata1=Xdata1,Xdata2=Xdata2,ncancorr=ncancorr,
                             Tau=Tau, CovStructure="Iden",standardize=TRUE,maxiteration=20,
                             thresh=0.0001)
}
}
\references{
Sandra E. Safo, Jeongyoun Ahn, Yongho Jeon, and Sungkyu Jung (2018),
 Sparse Generalized Eigenvalue Problem with Application to Canonical
 Correlation Analysis for Integrative Analysis of Methylation and Gene
 Expression Data. Biometrics
}
\seealso{
\code{\link{cvselpscca}}  \code{\link{cvtunerange}}
}
