% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlearnR-package.R
\docType{package}
\name{mvlearnR-package}
\alias{mvlearnR}
\alias{mvlearnR-package}
\title{mvlearnR: Multiview Learning Methods in R}
\description{
The mvlearnR package and accompanying Shiny App is intended for integrating data from multiple sources (e.g. genomics, proteomics, metabolomics). It is a compilation of various Multiview learning methods including SIDA and SIDANet (Sparse Integrative Discriminant Analysis for Multiview Structured Data), SELPCCA (Sparse Estimation via Linear Programming for canonical correlation analysis [CCA]), and SELP-Predict. The SIDA and SIDANet algorithms are for joint association and classification studies. The algorithms consider the overall association between multiview data, and the separation within each view when choosing discriminant vectors that are associated and optimally separate subjects. SIDANet incorporates prior structural information in joint association and classification studies. It uses the normalized Laplacian of a graph to smooth coefficients of predictor variables, thus encouraging selection of predictors that are connected and behave similarly. The SELPCCA method is an unsupervised method for associating two high dimensional data types. The algorithm obtains linear combinations of subsets of variables for each data type that contribute to overall dependency structure between the data types. SELP-Predict is a two-stage method for associating two views and predicting binary, continuous, poisson, or time-to-event data. Additional plotting and filtering functions such as variable importance plots, volcano plots, discriminant and correlation plots, relevance network and biplots are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.sandrasafo.com/software; https://multi-viewlearn.shinyapps.io/MultiView_Modeling/}
}

}
\keyword{internal}
