% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceMetrics.R
\name{PerformanceMetrics}
\alias{PerformanceMetrics}
\title{Performance Metrics}
\usage{
PerformanceMetrics(Y.pred, Y.test, family = "binomial", isPlot = TRUE)
}
\arguments{
\item{Y.pred}{A vector of predicted values. For SELPCCA, this is a vector of
predicted probabilities. For SIDA and SIDANet, this is a vector of predicted class.}

\item{Y.test}{A vector of observed test values.}

\item{family}{A string to denote the family for which  metrics should be provided.
Options are "gaussian", "binomial".}

\item{isPlot}{Boolean on whether or not to generate a plot. If family is binomial,
an AUC plot is generated. Currently applicable to SELPCCA. Set isPlot to FALSE
for SIDA and SIDANet. If isPlot is TRUE and predictions
are from SIDA and SIDANet, AUC estimate coincides with balanced accuracy.
If family is gaussian, a scatter plot of observed vs
predicted is generate.}
}
\value{
An output of performance metrics:
  \item{Metrics}{A table of estimated metrics}
}
\description{
Estimates performance metrics for a predicted model. Currently works
 for binary and continuous outcomes
}
\details{
For a binary outcome, we provide the following metrics:
         "Accuracy","Error rate","Sensitivity", "Specificity", "Matthews Correlation Coefficient (MCC)",
         "Balanced Accuracy","Balanced Error Rate", "F1 Score", "False.Discovery.Rate", and
         "Positive Predictive Value".

         For a  continuous outcome, we provide the following metrics:
         "Mean Squared Error","Root Mean.Squared Error","Relative Squared Error",
.         "Root Relative Squared.Error", "Root Absolute Error", "Mean Absolute Error".
}
\examples{
data(sidaData)
Xdata=sidaData[[1]]
Y=sidaData[[2]]
Xtestdata=sidaData[[3]]
Ytest=sidaData[[4]]
##---- call cross validation
 mycv=cvSIDA(Xdata,Y,withCov=FALSE,plotIt=FALSE, Xtestdata=Xtestdata,Ytest=Ytest,
             isParallel=TRUE,ncores=NULL,gridMethod='RandomSearch',
             AssignClassMethod='Joint',nfolds=5,ngrid=8,standardize=TRUE,
            maxiteration=20, weight=0.5,thresh=1e-03)
#check output
 test.error=mycv$sidaerror
 test.correlation=mycv$sidacorrelation
 optTau=mycv$optTau
 hatalpha=mycv$hatalpha
 #train metrics
 Y.pred=mycv$PredictedClass.train-1 #to get this in 0 and 1
 Y.train=Y-1 #to get this in 0 and 1
 train.metrics=PerformanceMetrics(Y.pred,Y.train,family='binomial',isPlot=FALSE)

 print(train.metrics)
 #obtain predicted class
 Y.pred=mycv$PredictedClass-1 #to get this in 0 and 1
 Ytest.in=Ytest-1 #to get this in 0 and 1
 test.metrics=PerformanceMetrics(Y.pred,Ytest.in,family='binomial',isPlot=FALSE)
 print(test.metrics)
}
\seealso{
\code{\link{cvSIDA}} \code{\link{selpscca.pred}} \code{\link{predict.SELPCCA}}
}
