% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selpData}
\alias{selpData}
\title{Data example for SELPscca}
\format{
A list with 7 elements:
\describe{
  \item{\code{Xdata1}}{A matrix of size \eqn{80 \times 200} for first dataset.
  Rows are samples and columns are variables.}
  \item{\code{Xdata2}}{A matrix of size \eqn{80 \times 150} for second dataset.
  Rows are samples and columns are variables.}
  \item{\code{Xtestdata1}}{A matrix of size \eqn{400 \times 200} for first dataset.
  Rows are samples and columns are variables.}
  \item{\code{Xtestdata2}}{A matrix of size \eqn{400 \times 150} for second dataset.
  Rows are samples and columns are variables.}
  \item{\code{TrueAlpha}}{The first canonical correlation vector for Xdata1.}
  \item{\code{TrueBeta}}{The first canonical correlation vector for Xdata2.}
  \item{\code{TrueCorr}}{The first canonical correlation coefficient.}
}
}
\usage{
selpData
}
\description{
Simulated data with one true canonical correlation
vectors for first and second datasets. The first 20 and 15
variables are nonzero (i.e., signal variables) in the first
canonical correlation vectors for the first and second datasets
respectively.
}
\references{
Sandra E. Safo, Jeongyoun Ahn, Yongho Jeon, and Sungkyu Jung (2018),
Sparse Generalized Eigenvalue Problem with Application to Canonical
Correlation Analysis for Integrative Analysis of Methylation and Gene
Expression Data. Biometrics
}
\keyword{datasets}
