% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sidaData}
\alias{sidaData}
\title{Data example for SIDA}
\format{
A list with 4 elements:
\describe{
  \item{\code{Xdata}}{A list with each entry containing two views
   of training data with dimension 160 × 2000 each. Rows are
   samples and columns are variables.}
  \item{\code{Y}}{160 × 1 vector of training class membership.
  There are two classes each with size 80.}
  \item{\code{Xtestdata}}{A list with each entry containing two views
   of testing data with dimension 320 × 2000 each. Rows are samples
    and columns are variables.}
  \item{\code{Ytest}}{ 320 × 1 vector of testing class membership.
  There are two classes each with size 160.}
}
}
\usage{
sidaData
}
\description{
Simulated data to demonstrate the use of SIDA.
}
\references{
Sandra E. Safo, Eun Jeong Min, and Lillian Haine (2019), Sparse Linear Discriminant
Analysis for Multi-view Structured Data, submitted.
}
\keyword{datasets}
