% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericPlots.R
\name{VarImportance}
\alias{VarImportance}
\title{Variable Importance Table}
\usage{
VarImportance(fit)
}
\arguments{
\item{fit}{the output from cvSIDA, cvSIDANet, or cvselpcca methods}
}
\value{
A dataframe of the absolute loadings for variables selected. The variables
are normalized to the variable with the largest weight.
}
\description{
returns Variable Importance tables for SIDA, SIDANet, or SELPCCA
}
\examples{
\dontrun{
##---- load SIDA data
data("sidaData")
Xdata <- sidaData[[1]]
Y <- sidaData[[2]]
Xtestdata <- sidaData[[3]]
Ytest <- sidaData[[4]]
##---- call cross validation
mycv=cvSIDA(Xdata,Y,withCov=FALSE,plotIt=FALSE, Xtestdata=Xtestdata,Ytest=Ytest, isParallel = FALSE)
##----  Obtain variable importance plot
VarImportance(mycv)
}
}
