% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(object)
}
\arguments{
\item{object}{the output from the filter.supervised() function}
}
\value{
A graph of the volcano plot
}
\description{
Wrapper function for volcano plots of the results after
supervised filtering.
}
\examples{
##---- read in data
data(COVID)

#make omics data numeric
Proteomics= apply(as.matrix(COVIDData[[1]]), 2, as.numeric)
RNASeq= apply(as.matrix(COVIDData[[2]]), 2, as.numeric)
Clinical= COVIDData[[3]]
X=list(Proteomics, RNASeq)
Y=Clinical$DiseaseStatus.Indicator

data.red=filter.supervised(X, Y,  method="t.test", padjust=TRUE,adjmethod="BH",
thresh=0.05,center=TRUE, scale=TRUE, Xtest=NULL)

##-----Volcano Plot of Result
volcanoPlot(data.red)
}
