% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringPlots.R
\name{umapPlot}
\alias{umapPlot}
\title{UMAP Plot}
\usage{
umapPlot(
  fit,
  useFilteredData = TRUE,
  usePrincipleComponents = TRUE,
  plotIt = TRUE
)
}
\arguments{
\item{fit}{the output from the filter_supervised() function}

\item{useFilteredData}{Boolean on whether to plot UMAP on filtered or original data.
Default is filtered data.}

\item{usePrincipleComponents}{boolean on whether to apply PCA first}

\item{plotIt}{boolean, whether to print the result or just return it (default = TRUE)}
}
\value{
A graph of the UMAP
}
\description{
Wrapper function to plot a UMAP of the results after
supervised filtering. See "umap" R package for more details on the
method.
}
\examples{
##---- read in data
data(sidaData)

Xdata=sidaData[[1]]
Y=sidaData[[2]]

data.red=filter_supervised(Xdata, Y,  method="t.test", padjust=TRUE, thresh=0.05,
                 center=FALSE, scale=FALSE, Xtest=NULL)

##-----Plot Result via UMAP
umapPlot(data.red)
}
