% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteringPlots.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(fit, plotIt = TRUE)
}
\arguments{
\item{fit}{the output from the filter_supervised() function}

\item{plotIt}{boolean, whether to print the result (TRUE) or just return it}
}
\value{
A graph of the volcano plot
}
\description{
Wrapper function for volcano plots of the results after
supervised filtering.
}
\examples{
##---- read in data
data(COVIDData)

#make omics data numeric
Proteomics= apply(as.matrix(COVIDData[[1]]), 2, as.numeric)
RNASeq= apply(as.matrix(COVIDData[[2]]), 2, as.numeric)
Clinical= COVIDData[[3]]
X=list(Proteomics, RNASeq)
Y=Clinical$DiseaseStatus.Indicator

data.red=filter_supervised(X, Y,  method="t.test", padjust=TRUE,adjmethod="BH",
thresh=0.05,center=TRUE, scale=TRUE, Xtest=NULL)

##-----Volcano Plot of Result
volcanoPlot(data.red)
}
